requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
anyio==3.5.0
    # via httpcore
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.0.12
    # via
    #   httpx
    #   requests
click==8.0.4
    # via typer
h11==0.12.0
    # via httpcore
httpcore==0.14.7
    # via httpx
httpx==0.22.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
jmespath==1.0.0
    # via -r requirements.in
msgpack==1.0.3
    # via scrapinghub
numpy==1.22.3
    # via pandas
pandas==1.4.1
    # via -r requirements.in
pyaml==21.10.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2021.3
    # via pandas
pyyaml==6.0
    # via pyaml
requests==2.27.1
    # via scrapinghub
retrying==1.3.3
    # via scrapinghub
rfc3986[idna2008]==1.5.0
    # via httpx
scrapinghub[msgpack]==2.4.0
    # via -r requirements.in
six==1.16.0
    # via
    #   python-dateutil
    #   retrying
    #   scrapinghub
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typer==0.4.0
    # via -r requirements.in
urllib3==1.26.9
    # via requests
